Moved frame GUI code to separate class files
